steam-crawler-master/requirements.txt
apache-airflow==3.0.3
numpy<2
pandas<2.2.0
Requests==2.32.4
SQLAlchemy<2.0
tqdm==4.67.1
urllib3==2.5.0
spacy==3.8.7
lingua-language-detector == 2.1.1
nltk == 3.9.1
polars == 1.30.0